requirements.txt
gensim==4.3.2
langchain_core==0.2.7
langchain_google_vertexai==1.0.5
langchain_openai==0.1.8
protobuf==5.27.1
PyYAML==6.0.1
spacy==3.7.5